requirements.txt
setuptools==59.5.0
numpy<2
torch
torchvision
openai-whisper
pyannote.audio
soundfile
speechbrain==0.5.16
python-dotenv